Update BombManController.cs
Fixed the target offset code. I had double the offset calculation and the logic for the offset was wrong. This was from when I was building the test code and just completely overlooked revising it. I thought it looked funny when I was making the video and now it's correct.

The other fix is for the forward (Jump2) & backward (Jump3) jump animation selections. I had only one case to cover the forward jump so the opposite was picking the backward jump when it shouldn't have been (he was forward jumping with the backward animation). I expanded the logic selection case and now it appears to be choosing the correct animation.