Stage Select Screen Stuck Fix
These are the changes for resolving the Stage Select Screen hanging after the next scene is called. If the Stage Select is the very first scene running it's fine however if it's called after any other scene then it hangs at its end. The problem is I forgot the add in the Start and Loop functions for the scene in the Game Manager.