Update MambuController.cs
Dynamically set Mambu's state. I use it to reset him back to his Closed state so his timer resets when he exits screen left and repositions to the right to re-enter. So if he was about to shoot when he leaves it doesn't happen right when he comes back into view.