Alternate Weapons Changes
Jumping onto the magnet beam doesn't seem to work anymore as you stop for an instant then fall through. I thought maybe it was the changing of the raycast distance of the boxcast but reverting it didn't make a difference. Adding a little extra padding by subtracting the box2d y extent from the test.

With the 4:3 resolutions version the transform find of the PlatformBeam now fails because the prefab will have the "_4x3" suffix. I changed it up a little and get the child  game object of the bulletShootPos then paranoid test its tag. Technically it should be a platform beam found because that's the currently selected weapon but anomalies can and do happen.

Last I made the CanUseWeaponAgain() method be public. As I have been building the BombMan stage and removing weapons when they're out of camera view, if a bomb for instance is removed, the event to let the player throw them again is never called from the explosion. So the weapon removal code will see if any weapons belong to the player and call CanUseWeaponAgain() so everything works as expected.