Resolution Scales in Scenes
I edited the Title Screen and Stage Select to make use of the resolution scale. They can adapt to the two scales I have set of 16:9 and 4:3. In the Title Screen the Press Any Key text will move lower when in 4:3. Before it was right up under the logo. In the Stage Select I made a lot of changes - the Canvas Scaler will update its Reference Resolution (pretty much zoom in for 4:3), new jump velocities for the enemies, bomb toss velocity, and Wily Ship start and end waypoints.